Version 4.3.1. Numerous performance inprovements through the use of SIMD acceleration in many functions.

+++ src/asum_simd_x86.h
@@ -1,15 +1,15 @@
 //------------------------------------------------------------------------------
 // Project:  arrayfunc
 // Module:   asum_simd_x86.h
-// Purpose:  Sum all the values in an array.
+// Purpose:  Calculate the asum of values in an array.
 //           This file provides an SIMD version of the functions.
 // Language: C
 // Date:     05-May-2017
-// Ver:      19-Jun-2018.
+// Ver:      04-Jun-2019.
 //
 //------------------------------------------------------------------------------
 //
-//   Copyright 2014 - 2018    Michael Griffin    <m12.griffin@gmail.com>
+//   Copyright 2014 - 2019    Michael Griffin    <m12.griffin@gmail.com>
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@
 //------------------------------------------------------------------------------
 
 
-float asum_float_simd(Py_ssize_t arraylen, float *data);
+double asum_float_simd(Py_ssize_t arraylen, float *data);
 double asum_double_simd(Py_ssize_t arraylen, double *data);
